Move BadgeUserList to its own file
